+++ NEWS.md
@@ -1,6 +1,7 @@
 # docorator development version
 Improvements
 * Add docx render engine `render_docx` (#97)
+* Convert all gt_groups to lists of gt_tbls for easier handling in render functions (#113)
 * update doc_pagenum() to support different render engines (#107)
 
 Bug Fixes